+++ NEWS.md
@@ -1,3 +1,22 @@
+# bcbioR 0.3.0
+
+* re-structure templates
+* Add text with best practices
+* Reproducibility:
+  * test data for RNAseq, singlecell, CHIPseq
+* Base project:
+  * Guidelines to create repo easily
+  * Example Rmd with headers and aesthetics
+* RNASEQ
+  * Use provenance for FA in DE report
+  * Support multiple comparisons
+* New templates:
+  * methylation - draft
+  * singcell cell QC and Inegration - stable
+  * scQC shiny app - stable
+  * chipseq QC and Diffbind - beta
+  * COSMX - draft
+
 # bcbioR 0.1.3
 
 * fix duplicated gene names